inspect_payload returns matmul shapes, move to mlir utils
